Rename date.rb to cppdate.rb
Other changes:

- Options have been removed because this is a new formula
- make -j; make install have been replaced by make install
- Test have been change to test that the library is linking properly
